Adding the small Platform assets to the tool
This merge I have added the small assets that I made for the platforms and gave some colliders I need to check this is ok the idea was for gameplay but need to check with team. I have now added some assets to the platform for each level as well. I have one more section to add to the tool and that is the special platforms once made.